requirements.txt
Flask==3.0.0
flask-cors==4.0.0
adbutils==2.12.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0
mutagen==1.47.0

reportlab==4.0.7
